Avoid removable and loop devices in disk stats
... and also be aware that in future there will be 4k devices
as well.

+++ linux/diskmeter.h
@@ -11,8 +11,7 @@
 #include <map>
 #include <string>
 
-typedef std::map<std::string, unsigned long> diskmap;
-
+typedef std::map<std::string, unsigned long long> diskmap;
 
 class DiskMeter : public FieldMeterGraph
     {
